use stack justify to make content flexible fixed
