+++ packages/app/src/app/pages/Sandbox/Editor/Workspace/screens/ProjectInfo/Config.tsx
@@ -1,14 +1,13 @@
 import React, { MouseEvent } from 'react';
 import getTemplateDefinition from '@codesandbox/common/lib/templates';
 import { useOvermind } from 'app/overmind';
-import { css } from '@styled-system/css';
-import { Button, Element, Stack } from '@codesandbox/components';
+import { Button, Element } from '@codesandbox/components';
 
 export const Config = () => {
   const {
     actions: {
       modalOpened,
-      workspace: { addedTemplate, deleteTemplate },
+      workspace: { addedTemplate },
     },
     state: {
       user,
@@ -20,16 +19,6 @@ export const Config = () => {
       },
     },
   } = useOvermind();
-
-  const onDelete = (event: MouseEvent<HTMLButtonElement>) => {
-    event.preventDefault();
-
-    if (customTemplate) {
-      deleteTemplate();
-    } else {
-      modalOpened({ modal: 'deleteSandbox' });
-    }
-  };
 
   const onCreateTemplate = (event: MouseEvent<HTMLButtonElement>) => {
     event.preventDefault();
@@ -60,26 +49,6 @@ export const Config = () => {
           </Button>
         </Element>
       )}
-
-      <Stack
-        justify="center"
-        css={css({
-          bottom: 3,
-          position: 'absolute',
-          width: '100%',
-        })}
-      >
-        <Button
-          // @ts-ignore
-          onClick={onDelete}
-          variant="link"
-          css={css({
-            ':hover,:focus': { color: 'dangerButton.background' },
-          })}
-        >
-          {`Delete ${customTemplate ? `Template` : `Sandbox`}`}
-        </Button>
-      </Stack>
     </>
   );
 };
