+++ packages/app/src/app/pages/Sandbox/Editor/Workspace/screens/ProjectInfo/index.tsx
@@ -1,12 +1,17 @@
 import React from 'react';
+import { Stack } from '@codesandbox/components';
 import { Summary } from './Summary';
 import { Privacy } from './Privacy';
 import { Config } from './Config';
+import { Delete } from './Delete';
 
 export const ProjectInfo = () => (
-  <>
-    <Summary />
-    <Privacy />
-    <Config />
-  </>
+  <Stack direction="vertical" justify="space-between" css={{ height: '100%' }}>
+    <div>
+      <Summary />
+      <Privacy />
+      <Config />
+    </div>
+    <Delete />
+  </Stack>
 );
